refactor room board creation
